feat: Implement HomeClient component for enhanced HomePage functionality with loading state
