+++ apps/ticket-admin/src/app/_clientBoundary/HomeClient/index.module.scss
@@ -12,7 +12,7 @@
 .title {
   margin-bottom: 1rem;
 
-  color: #333;
+  color: #fff;
   font-weight: bold;
   font-size: 3rem;
 }
